--- a/scripts/index.js
+++ b/scripts/index.js
@@ -17,6 +17,10 @@ var BABEL_MINI_REPL = (function() {
     compileCode(inEditor, outEditor);
   }, 1000);
 
+  function isMobile() {
+    return window.screen.width < 760;
+  }
+
   function setupEditor(id, readOnly) {
     var editor = ace.edit(id);
 
@@ -130,9 +134,9 @@ var BABEL_MINI_REPL = (function() {
   return {
     start: function() {
       // don't init editor on mobile devices
-      if (window.screen.width < 760) return;
+      if (isMobile()) return;
 
-      $('.hero-repl').attr('hidden', false);
+      $('.hero-repl').prop('hidden', false);
 
       inEditor = setupEditor('hero-repl-in', true);