Hey guys :wave:

Tracing site a bit I found a place for opportunity.
Using [pwmetrics](https://github.com/paulirish/pwmetrics) were made some measurements and noticed that Time to Interactive happens too late after First Meaningful Paint.

![Alt text](https://monosnap.com/file/vOBOnzjxq1Kyc80ZZOr6mlQH2HUOYC.png)

Also, I traced stuff on a Mi Redmi 3 phone with 3G network throttling.

Here the trace - https://chromedevtools.github.io/timeline-viewer/?loadTimelineFromURL=drive://0B0c67TI7mLzEN0VMYkJhdW1UZjQ

`FCP` ~ `3,1sec`
`TTI` ~ `6,8sec`

Not nice :(

----
~After digging in, it turned out that there were no `defer` attributes for scripts, moreover no minified css and uglified js where used.
For minifying assets [jekyll-assets](https://github.com/jekyll/jekyll-assets) was used (it was playing with magic).~
So after adding defer

![Alt text](https://monosnap.com/file/iYL62FjzJskoe6VBTdv0DwQdYX6n5k.png)

`FCP` ~ `3sec` (saved 100ms)
`TTI` ~ `6,7sec` (saved 100ms)

Also, trace with improvements for the phone - https://chromedevtools.github.io/timeline-viewer/?loadTimelineFromURL=drive://0B0c67TI7mLzETHFMd1lHNTRxalk

---

Nicer but `TTI` is still not good :(

Ok, what's wrong?

A lot of problems were caused by [ace.js](https://github.com/ajaxorg/ace)
![Alt text](https://monosnap.com/file/b2ww703qc2co4wxGoFA1pD9FFo7WxF.png)

So I suggest just don't show fancy editor stuff for mobile devices on the main page. Users on a mobile just read docs, I doubt that they wanna play with editor online. I would not to :)

---

So...... :tropical_fish:

:tada:
![Alt text](https://monosnap.com/file/P7wIWIgrNlpUz05EEBYq2IiYOtTQFL.png)

`FCP` ~ `3sec` (saved 100ms)
`TTI` ~ `3,3sec` (saved 2,5sec)

And timeline if someone interested - https://chromedevtools.github.io/timeline-viewer/?loadTimelineFromURL=drive://0B0c67TI7mLzEY3pfS0M2NEw2SUU

> Note: these timelines are true for using `defer` with `jekyll assets` together. 

UPD: fix spelling.
UPD2: removed `jekyll-assets` stuff