@existentialism 2017-04-16
@denar90 thanks for the analysis and the PR!

Definitely agree re: mini-repl on mobile, I had made a point to come back and disable it on mobile, but never got to it.

One complication with some of the changes (and why we haven't already done most of this) is due to our hosting on GitHub Pages, we can't actually use `jekyll-assets` (ref: https://github.com/github/pages-gem/issues/189).

We _could_ build on CI and deploying the built files to GH pages, or change hosts altogether, but those are probably more involved discussions.