@xtuc 2017-04-16
This sounds awesome, thanks for your PR :+1:

@babel-bot could help us with deploying the website (i'm not sure how yet).