@denar90 2017-04-17
Thanks guys for looking into this. How do you ship the site? E.g. for shipping @marionettejs site we use CI build approach - https://github.com/marionettejs/marionettejs.com/blob/master/travis-runner.sh
The same stuff can be helpful here
Kinda
```sh
if [ "$TRAVIS_BRANCH" = "master" ] && [ "$TRAVIS_PULL_REQUEST" = "false" ]
then
  echo "Deploying!"
  bower i
  npm run build
  cp CNAME _site
  cd _site
  git config --global user.email "travis@babel.io"
  git config --global user.name "TRAVIS-CI"
  git init
  git add .
  git commit -m "deploy"
  git push --force --quiet "https://${GH_TOKEN}@babel/babel.github.io.git" master:gh-pages > /dev/null 2>&1
else
  npm run lint
fi
```