@denar90 2017-04-17
But also I have one coplaine about `Gemfile.lock` changes - https://github.com/babel/babel.github.io/pull/1234/files#diff-e79a60dc6b85309ae70a6ea8261eaf95
Is it ok, if we get rid of all these deps?