@xtuc 2017-04-17
Currently we sync manually babel/babel and babel/babel.github.io, this is not the right way for doing it.

I will test your script, thanks :+1: 

About the gems being removed, I don't see any differences.