@existentialism 2017-04-17
I don't think it's a bad idea to use the github pages gem as it let's us mimic the hosting environment. 

Ultimately, I'm fine if we want to switch to travis building/deploying the site... it would also let us run some custom gems to clean up some hackery, etc.