@denar90 2017-04-17
What the plan guys, keep this PR until deploy stuff will be configured?