@xtuc 2017-04-18
The deployment stuff is a bit more complicated here because we rely on git submodules. We use them to publish documentation from Babel on our website.

This looks good to me.

@existentialism are you fine with the gem changes?

@denar90 Could we land this without the gem changes?