@denar90 2017-04-18
>@denar90 Could we land this without the gem changes?

@xtuc `jekyll-access` is required to be used because it changes the name to minified files..
But, guys, if I understood you correctly, you land site manually, which means that you run all commands and deploy `_site` folder. If so, then nothing changed for your process, because assets generated by `jekyll-access` are put into `_site` folder.