@existentialism 2017-04-18
My point was, is that we _can't_ land the gem changes without changing how we deploy.

@denar90 we don't run commands and deploy `_site` folder... we currently rely on GitHub's built-in build/deploy (hence using the `github-pages` gem to ensure dev env matches)

We can land adding `defer` attrs to script tags and hiding mini-repl on mobile, and then figure out changing deployment to add minification, etc in another PR?