@xtuc 2017-04-18
I agree with @existentialism, well said.