@denar90 2017-04-18
Sounds reasonable guys. I'm totally 👍 for doing it this way.
I'll drop `jekyll-access` commits, so we will be able to land this PR :)