@denar90 2017-04-18
I've updated audit comment and removed stuff for `jekyll-assets`.