@existentialism 2017-04-20
@denar90 thanks again for the awesome analysis.. I created https://github.com/babel/babel.github.io/issues/1238 to continue discussion (would love your help/input here too!)